requirements.txt
Flask==3.0.0
Pillow>=10.0.1
Werkzeug==3.0.1
google-genai==0.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
